+++ build_runner_core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Removed unused dev_deps: `test_process`.
 - Bug fix: fix incremental build after build with import of missing package.
+- Testing: read build configs using `AssetReader` so they're easier to test.
 
 ## 9.2.0-dev.1